+++ src/kohlrahbi/ahbtable/ahbsubtable.py
@@ -9,7 +9,7 @@ from docx.table import Table as DocxTable  # type:ignore[import]
 from docx.table import _Cell  # type:ignore[import]
 from pydantic import BaseModel, ConfigDict
 
-from kohlrahbi.ahb.ahbtablerow import AhbTableRow
+from kohlrahbi.ahbtable.ahbtablerow import AhbTableRow
 from kohlrahbi.row_type_checker import RowType, get_row_type
 from kohlrahbi.seed import Seed
