+++ src/kohlrahbi/ahbtable/ahbtable.py
@@ -10,7 +10,7 @@ from maus.edifact import get_format_of_pruefidentifikator
 from more_itertools import peekable
 from pydantic import BaseModel, ConfigDict
 
-from kohlrahbi.ahb.ahbsubtable import AhbSubTable
+from kohlrahbi.ahbtable.ahbsubtable import AhbSubTable
 from kohlrahbi.logger import logger
 from kohlrahbi.table_header import PruefiMetaData
